+++ AR41DX/Engine/Include/Resource/ResourceManager.h
@@ -164,7 +164,7 @@ public:	// ============================ Sound ================================
 	bool CreateSoundChannel(const std::string& Name);
 	bool LoadSound(const std::string& GroupName, const std::string& Name,
 		bool Loop, const char* FileName, const std::string& PathName = SOUND_PATH);
-	bool SetVolume(int Volume);
+	bool SetVolume(float Volume);
 	bool SetVolume(const std::string& GroupName, int Volume);
 	bool SoundPlay(const std::string& Name);
 	bool SoundStop(const std::string& Name);
